[finance_report_vision] refactor(#1670): fold prompts/reconciliation.py into the reconciliation package
The last non-fx app-boundary edge: reconciliation/extension/scoring.py's AI
semantic-match prompt lived in src/prompts/reconciliation.py, unowned by any
package. It has a single production consumer (scoring.py itself) and zero
dependencies (a pure prompt template + string-formatting function), so it
belongs in reconciliation/base/ per the base/extension/data convergence rule.

- Move prompts/reconciliation.py -> reconciliation/base/prompts.py.
- Publish RECONCILIATION_SEMANTIC_PROMPT/build_reconciliation_prompt on
  reconciliation's root + common/reconciliation/contract.py's interface.
- Repoint scoring.py's local import and the one stray test consumer.
- app-boundary-baseline.json: 6 -> 5 edges (only the 4 fx.py edges, owned by
  #1610 P2 + #1641, and the models/_registry.py ORM-registration edge remain).
