+++ apps/backend/src/reconciliation/__init__.py
@@ -5,9 +5,11 @@ from __future__ import annotations
 from src.reconciliation.base import (
     DEFAULT_CONFIG,
     MAX_COMBINATION_CANDIDATES,
+    RECONCILIATION_SEMANTIC_PROMPT,
     MatchCandidate,
     ReconciliationConfig,
     _candidate_is_better,
+    build_reconciliation_prompt,
     entry_bank_side_amount,
     entry_total_amount,
     is_entry_balanced,
@@ -80,6 +82,7 @@ __all__ = [
     "FxTransferError",
     "MAX_COMBINATION_CANDIDATES",
     "MatchCandidate",
+    "RECONCILIATION_SEMANTIC_PROMPT",
     "ReconciliationConfig",
     "ReconciliationStats",
     "TransferLeg",
@@ -95,6 +98,7 @@ __all__ = [
     "auto_accept",
     "batch_accept",
     "build_many_to_one_groups",
+    "build_reconciliation_prompt",
     "calculate_match_score",
     "classify_internal_transfer",
     "detect_anomalies",
