+++ apps/backend/src/reconciliation/base/__init__.py
@@ -12,16 +12,22 @@ from src.reconciliation.base.config import (
     is_entry_balanced,
     load_reconciliation_config,
 )
+from src.reconciliation.base.prompts import (
+    RECONCILIATION_SEMANTIC_PROMPT,
+    build_reconciliation_prompt,
+)
 from src.reconciliation.base.repository import ReconciliationRepository
 
 __all__ = [
     "DEFAULT_CONFIG",
     "MAX_COMBINATION_CANDIDATES",
     "MatchCandidate",
+    "RECONCILIATION_SEMANTIC_PROMPT",
     "ReconciliationConfig",
     "ReconciliationRepository",
     "_candidate_is_better",
     "_candidate_source_rank",
+    "build_reconciliation_prompt",
     "entry_bank_side_amount",
     "entry_total_amount",
     "is_entry_balanced",
