+++ apps/backend/src/reconciliation/extension/scoring.py
@@ -243,7 +243,7 @@ async def ai_semantic_score(
     import json
 
     from src.llm import AIStreamError, accumulate_stream, stream_ai_json
-    from src.prompts.reconciliation import build_reconciliation_prompt
+    from src.reconciliation.base.prompts import build_reconciliation_prompt
 
     prompt = build_reconciliation_prompt(
         txn_description=txn_description,
