+++ common/reconciliation/contract.py
@@ -140,6 +140,7 @@ CONTRACT = PackageContract(
         "FxTransferError",
         "MAX_COMBINATION_CANDIDATES",
         "MatchCandidate",
+        "RECONCILIATION_SEMANTIC_PROMPT",
         "ReconciliationConfig",
         "ReconciliationStats",
         "TransferLeg",
@@ -155,6 +156,7 @@ CONTRACT = PackageContract(
         "auto_accept",
         "batch_accept",
         "build_many_to_one_groups",
+        "build_reconciliation_prompt",
         "calculate_match_score",
         "classify_internal_transfer",
         "detect_anomalies",
